put everything offchain voting related in one js lib file
