+++ utils/offchain_voting.js
@@ -212,6 +212,87 @@ function bufDedup(buffers) {
   })
 }
 
+async function addVote(votes, snapshotRoot, daoAddress, proposalId, account, memberWeight, voteYes) {
+  const proposalHash = sha3(web3.eth.abi.encodeParameters(
+    ['bytes32', 'address', 'uint256'], 
+    [snapshotRoot, daoAddress, proposalId]));
+  const vote = {
+    address : account.toString(),
+    weight: memberWeight,
+    signature : await generateVote(account, proposalHash, voteYes),
+    voteResult : voteYes ? '1' : '2'
+  };
+  votes.push(vote);
+  return votes;
+}
+
+async function generateVote(account, proposalRoot, voteYes) {
+  const voteHash = sha3(web3.eth.abi.encodeParameters( ['bytes32', 'uint256'], [proposalRoot, voteYes ? "1" : "2"]));
+  return await web3.eth.sign(voteHash, account);
+}
+
+async function prepareSnapshot(dao, member, accounts) {
+  const shares = await Promise.all(accounts.map(async (account) => {
+    const nbShares = await member.nbShares(dao.address, account);
+    return {account, nbShares};
+  }));
+
+  const cleanShares =  shares.filter(({nbShares}) => nbShares.toString() !== '0');
+
+  const elements = cleanShares
+  .sort((a, b) => a.account > b.account)
+  .map(({account, nbShares}) => sha3(web3.eth.abi.encodeParameters(['address', 'uint256'], [account, nbShares.toString()])));
+  const weights = cleanShares.reduce((o, elem) => Object.assign(o, {[elem.account]: elem.nbShares}), {});
+
+  return {snapshotTree: MerkleTree(elements, true), weights};
+}
+
+function buildVoteLeafHashForMerkleTreeData(leaf) {
+  return web3.eth.abi.encodeParameters(
+    ['address', 'uint256', 'bytes', 'uint256', 'uint256'], 
+    [leaf.address, leaf.weight.toString(), leaf.vote.toString(), leaf.nbYes.toString(), leaf.nbNo.toString()]);
+}
+
+function buildVoteLeafHashForMerkleTree(leaf) {
+  return sha3(buildVoteLeafHashForMerkleTreeData(leaf));
+}
+
+function prepareVoteResult(votes, weights) {
+  const sortedVotes = votes.sort((a,b) => a.address > b.address);
+  const leaves = sortedVotes.map((vote) => {
+    return {
+      address: vote.address,
+      weight: weights[vote.address],
+      vote: vote.signature,
+      voteResult: vote.voteResult
+    }
+  });
+
+  leaves.forEach((leaf, idx) => {
+    leaf.nbYes = leaf.voteResult === 1 ? 1 : 0;
+    leaf.nbNo = leaf.voteResult !== 1 ? 1 : 0;
+    
+    if(idx > 0) {
+      const previousLeaf = leaves[idx - 1];
+      leaf.nbYes = leaf.nbYes + previousLeaf.nbYes;
+      leaf.nbNo = leaf.nbNo + previousLeaf.nbNo;
+    }
+    
+  });
+
+  leaves.forEach((leaf) => {
+    leaf.nbYes = leaf.voteResult === 1 ? 1 : 0;
+    leaf.nbNo = leaf.voteResult !== 1 ? 1 : 0;
+  });
+
+  const voteResultTree = new MerkleTree(leaves.map(vote => buildVoteLeafHashForMerkleTree(vote), true));
+
+  return {voteResultTree, votes: leaves};
+}
+
 Object.assign(exports, {
-    MerkleTree
+    MerkleTree,
+    prepareSnapshot,
+    addVote,
+    prepareVoteResult
   })
